backend/requirements.txt
# Studio Backend - Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment
python-dotenv>=1.0.0

# AI APIs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Settings
pydantic-settings>=2.0.0

# HTTP Client
httpx>=0.26.0

# File Upload
python-multipart>=0.0.6

# Spaced Repetition
fsrs>=3.0.0

# Supabase
supabase>=2.3.0

# PDF Processing
PyPDF2>=3.0.0

# Async
anyio>=4.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
